ci: add automated release workflow
The new release.yml triggers the release automation. It increases the
version.json if needed and it should create a CHANGELOG.md file. We also
added a git hook to help enforcing the conventional commits specs. A
message for this is added into the CMakeLists.txt.  All this is
documented in the README.md.

Refs: #106